+++ packages/core/src/components/Table/ColumnConfiguration/ColumnConfiguration.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Checkbox from '../../Checkbox';
 import CheckboxGroup from '../../CheckboxGroup';
-import { ColumnConfig } from '../types';
+import { TableColumnConfig } from '../types';
 import { createOptions, createValues, updateColumns, updateConfig } from './helpers';
 import { Props } from './types';
 
@@ -14,10 +14,11 @@ const ColumnConfiguration: React.FC<Props> = React.memo(({ columns, onChange }) 
             onChange(newColumns);
         };
 
-    const checkBoxes = (configs: ColumnConfig[]) =>
+    const checkBoxes = (configs: TableColumnConfig[]) =>
         configs.map(config =>
             config.children ? (
                 <CheckboxGroup
+                    fullWidth
                     showSelectAll
                     key={config.field}
                     label={config.title}
@@ -27,7 +28,13 @@ const ColumnConfiguration: React.FC<Props> = React.memo(({ columns, onChange }) 
                     options={createOptions(config.children, config.field)}
                 />
             ) : (
-                <Checkbox key={config.field} label={config.title} checked={config.hide} onChange={handleCheckboxClick(config.field)} />
+                <Checkbox
+                    fullWidth
+                    key={config.field}
+                    label={config.title}
+                    checked={config.hidden}
+                    onChange={handleCheckboxClick(config.field)}
+                />
             )
         );