+++ packages/core/src/components/Table/docs/columns.tsx
@@ -1,19 +1,19 @@
 import { styled } from '@medly-components/utils';
 import React from 'react';
 import Chip from '../../Chip';
-import { ColumnConfig } from '../types';
+import { TableColumnConfig } from '../types';
 
 const StyledDiv = styled('div')`
         display: flex;
         flex-direction: column;
     `,
-    FavoriteColor: ColumnConfig['component'] = ({ data: colors, disabled }) => {
+    FavoriteColor: TableColumnConfig['component'] = ({ data: colors, disabled }) => {
         const chips = colors.split(',').map((color: string) => <Chip key={color} label={color} disabled={disabled} />);
         return <StyledDiv>{chips}</StyledDiv>;
     };
 
-const columns: ColumnConfig[] = [
-    { title: 'Name', field: 'name', formatter: 'text-short', frozen: true, fitContent: true },
+const columns: TableColumnConfig[] = [
+    { title: 'Name', field: 'name', formatter: 'text-short', sortable: true, fitContent: true },
     {
         title: 'Marks',
         field: 'marks',
@@ -23,25 +23,18 @@ const columns: ColumnConfig[] = [
                 title: 'Maths',
                 field: 'maths',
                 formatter: 'numeric',
-                sort: true,
+                sortable: true,
                 children: [
-                    { title: 'Algebra', field: 'algebra', formatter: 'numeric', sort: true },
-                    { title: 'Trigonometry', field: 'trigonometry', formatter: 'numeric', sort: true }
+                    { title: 'Algebra', field: 'algebra', formatter: 'numeric', sortable: true },
+                    { title: 'Trigonometry', field: 'trigonometry', formatter: 'numeric', sortable: true }
                 ]
             },
-            { title: 'Science', field: 'science', formatter: 'numeric', sort: true }
+            { title: 'Science', field: 'science', formatter: 'numeric', sortable: true }
         ]
     },
-    { title: 'Age', field: 'age', formatter: 'numeric', sort: true },
-    { title: 'Passed', field: 'isPassed', formatter: 'boolean', sort: true },
-    {
-        title: 'Favorite Color',
-        field: 'color',
-        formatter: 'react-component',
-        component: FavoriteColor,
-        sort: true
-    },
-    { title: 'Rating', field: 'rating', formatter: 'numeric', sort: true }
+    { title: 'Age', field: 'age', formatter: 'numeric', sortable: true },
+    { title: 'Passed', field: 'isPassed', formatter: 'boolean', sortable: true },
+    { title: 'Rating', field: 'rating', formatter: 'numeric', sortable: true }
 ];
 
 export default columns;